requirements.txt
pandas==1.5.0
geopandas==0.11.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
# shapely
python-dotenv==0.21.0
